Renames and fix-ups for readability.
