+++ src/test/java/org/humancellatlas/ingest/schemas/SchemaScraperTest.java
@@ -1,7 +1,7 @@
 package org.humancellatlas.ingest.schemas;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
-import org.humancellatlas.ingest.core.MigrationConfiguration;
+import org.humancellatlas.ingest.config.MigrationConfiguration;
 import org.humancellatlas.ingest.schemas.schemascraper.SchemaScraper;
 import org.humancellatlas.ingest.schemas.schemascraper.impl.S3BucketSchemaScraper;
